backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wta-backend"
version = "0.1.0"
description = "Milestone 1 backend for the WTA fitness tracker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "bcrypt>=4.0.1,<4.1.0",
  "fastapi>=0.115.0,<1.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
